feat: move storage from SQLite to PostgreSQL
Run PostgreSQL 17 in Docker via docker-compose on host port 5433, since
5432 is taken by another container on this machine. The data volume is
mounted at /var/lib/postgresql so it survives container removal.

The schema now uses native uuid and timestamptz columns, which removes
the to_string() binds and try_from attribute the TEXT-backed ids needed
under SQLite. Queries switch from ? to $n placeholders.

limit and offset become i64 because Postgres has no unsigned integers.
That loses the free rejection of negative values, so offset gets an
explicit range(min = 0) and ?limit=-1 is now a 422 rather than a 400;
test.fish is updated to match and gains an ?offset=-1 check.

+++ src/models/task.rs
@@ -5,7 +5,7 @@ use validator::Validate;
 
 #[derive(Debug, Serialize, Deserialize, Clone, sqlx::Type)]
 #[serde(rename_all = "snake_case")]
-#[sqlx(type_name = "task_status", rename_all = "snake_case")]
+#[sqlx(type_name = "TEXT", rename_all = "snake_case")]
 pub enum TaskStatus {
     Todo,
     InProgress,
@@ -13,7 +13,6 @@ pub enum TaskStatus {
 }
 #[derive(Debug, Serialize, Clone, sqlx::FromRow, Validate)]
 pub struct Task {
-    #[sqlx(try_from = "String")]
     pub(crate) id: Uuid,
     #[validate(length(min = 1, max = 200))]
     pub(crate) title: String,
@@ -63,8 +62,7 @@ pub struct UpdateTask {
     pub(crate) description: Option<Option<String>>,
     pub(crate) status: Option<TaskStatus>,
 }
-
-fn default_limit() -> u32 {
+fn default_limit() -> i64 {
     20
 }
 
@@ -73,9 +71,10 @@ pub struct TaskQuery {
     pub status: Option<TaskStatus>,
     #[serde(default = "default_limit")]
     #[validate(range(min = 1, max = 100, message = "limit must be 1-100"))]
-    pub limit: u32,
+    pub limit: i64,
     #[serde(default)]
-    pub offset: u32,
+    #[validate(range(min = 0, message = "offset must not be negative"))]
+    pub offset: i64,
 }
 
 #[cfg(test)]
